debugged removing tiles

+++ SFMLEngine/src/SFMLEngine/Application.cpp
@@ -192,7 +192,7 @@ namespace SFMLEngine
             
             if (m_DisplayDebug)
             {
-                fps = round(1 / ts);
+                fps = (int)round(1 / ts);
                 m_DebugInfo.push_back("FPS: " + std::to_string(fps));
             }
